trying

+++ bs/src/schemas.py
@@ -49,29 +49,27 @@ class ModelRatingOut(BaseModel):
     category: str
 
     net_score: float
-    net_score_latency: int
+    net_score_latency: float
     ramp_up_time: float
-    ramp_up_time_latency: int
+    ramp_up_time_latency: float
     bus_factor: float
-    bus_factor_latency: int
+    bus_factor_latency: float
     performance_claims: float
-    performance_claims_latency: int
+    performance_claims_latency: float
     license: float
-    license_latency: int
+    license_latency: float
     dataset_and_code_score: float
-    dataset_and_code_score_latency: int
+    dataset_and_code_score_latency: float
     dataset_quality: float
-    dataset_quality_latency: int
+    dataset_quality_latency: float
     code_quality: float
-    code_quality_latency: int
-
+    code_quality_latency: float
     reproducibility: float
-    reproducibility_latency: int
+    reproducibility_latency: float
     reviewedness: float
-    reviewedness_latency: int
+    reviewedness_latency: float
     tree_score: float
-    tree_score_latency: int
-
+    tree_score_latency: float
     size_score: SizeScoreOut
-    size_score_latency: int
+    size_score_latency: float
